Change popup redraw
Now, when the last item is removed from the dump, it looks like it was
cleared eventhough an array is still in the storage.

+++ popup/load.content.js
@@ -5,31 +5,34 @@ function deleteItem(event) {
   });
 }
 
-browser.storage.local.get('urls').then(obj => {
-  if (obj.urls) {
-    obj.urls.forEach((item, index) => {
-      const listItem = document.createElement('p');
-      const itemLink = document.createElement('a');
-      itemLink.href = item.url;
-      itemLink.innerHTML = item.title;
-      const itemDelete = document.createElement('button');
-      itemDelete.dataset.index = index;
-      itemDelete.onclick = deleteItem;
-      const deleteImage = document.createElement('img');
-      deleteImage.src = browser.extension.getURL('icons/trash-48.png');
-      itemDelete.appendChild(deleteImage);
+function loadDump() {
+  document.querySelector('#popup-content').innerHTML = '';
+  browser.storage.local.get('urls').then(obj => {
+    if (obj.urls && obj.urls.length !== 0) {
+      obj.urls.forEach((item, index) => {
+        const listItem = document.createElement('p');
+        const itemLink = document.createElement('a');
+        itemLink.href = item.url;
+        itemLink.innerHTML = item.title;
+        const itemDelete = document.createElement('button');
+        itemDelete.dataset.index = index;
+        itemDelete.onclick = deleteItem;
+        const deleteImage = document.createElement('img');
+        deleteImage.src = browser.extension.getURL('icons/trash-48.png');
+        itemDelete.appendChild(deleteImage);
 
-      listItem.appendChild(itemDelete);
-      listItem.appendChild(itemLink);
+        listItem.appendChild(itemDelete);
+        listItem.appendChild(itemLink);
 
-      document.querySelector('#popup-content').appendChild(listItem);
-    });
-  } else {
-    const emptyItem = document.createElement('p');
-    emptyItem.innerHTML = 'The dump is empty!';
-    document.querySelector('#popup-content').appendChild(emptyItem);
-  }
-});
+        document.querySelector('#popup-content').appendChild(listItem);
+      });
+    } else {
+      const emptyItem = document.createElement('p');
+      emptyItem.innerHTML = 'The dump is empty!';
+      document.querySelector('#popup-content').appendChild(emptyItem);
+    }
+  });
+}
 
 document.querySelectorAll('.download').forEach(item => {
   item.addEventListener('click', (event) => {
@@ -56,7 +59,7 @@ document.querySelector('.dropdown-toggle').addEventListener('click', (event) => 
 function handleMessage(message) {
   switch (message.action) {
     case 'reload':
-      window.location.reload(true);
+      loadDump();
       break;
     default:
       // Do nothing on purpose
@@ -64,3 +67,4 @@ function handleMessage(message) {
 }
 
 browser.runtime.onMessage.addListener(handleMessage);
+loadDump();
